add POST /tags/
