+++ app/tags/tests/views/test.py
@@ -1,0 +1,20 @@
+from app.base.tests.fakers import fake
+from app.base.tests.views.base import BaseViewTest
+from app.tags.models import Tag
+from app.tags.tests.factories import CategoryFactory
+
+
+class TagsTest(BaseViewTest):
+    path = '/tags/'
+
+    def test_post(self):
+        category = CategoryFactory()
+        self._test(
+            'post',
+            data={
+                'shortcut': fake.first_name(),
+                'title': fake.first_name(),
+                'categories': [category.id],
+            },
+        )
+        self.assert_model(Tag, {})
